Force an update of the extruder configs when both extruders are detected
The root cause of this extruder combo mismatch bug seems to be to not have
the latest extruder config updates from both extruders leading to using a
stale update for the first installed extruder. Since the condition for the
combo mismatch check strictly looks for the approval of both the extruders,
we just force an extruder config update from kaiten when both extruders are
attached or when both are attached and one of them is removed.

BW-5894
https://ultimaker.atlassian.net/browse/BW-5894