Don't persist unused settings keys
If we change the naming of settings keys, or remove settings keys, or if someone manually edits a settings key with a typo into their persisent settings, we would like to remove that unused key.  

This code change will not necessarily remove the settings key immediately, but it will eventually remove it the first time we need to write a settings change back to disk.