Update qsTr() usage
BW-5898
http://ultimaker.atlassian.net/browse/BW-5898

We want qsTr() wrapping anything we display to the user that we might
possibly want to translate.  Conversely we don't want it wrapping any
strings that we don't want to translate, like a URL that we display
to the user or any string that is used on the backend.  (We also don't
want placeholder text like "HEADLINE" here, but I only removed these
when we didn't already have a translation.)

One other important detail about making strings translatable is that
you really don't want to be using logic to build sentence structure.
So some code is refactored here to list out each full sentence that
we might construct, though I also just removed some inaccessible code
instead of rewriting it.

I also skipped updating the spool info page since it isn't user
accessible, but I also didn't delete it.