Add general help state to help popup
BW-5914
http://ultimaker.atlassian.net/browse/BW-5914

We were missing a popup for the Unload material (?) screen. I added the general help popup as per the spec and connected it during unload material rewind spool page.

I am not sure if it's possible to get to the unload material page through the FRE, but I added a check, but let me know if that is incorrect. The general help qr code points to the support page. I also changed the text for the base state to say "HELP" instead of "METHOD COMPATIBILITY" as well.