Support for magma top lid removal
The main blocker here is that the attach extrders process has a step to
wait for the lid to be opened, which just waits for the lid open error
to be reported.  Additionally we add messaging for error 1001 for the
user to check the lid.

I did not make either of these changes magma specific because they seem
like good ideas in general.  There isn't any really good reason to block
a first time user from skipping past the open lid step, and anyone who
goes through the FRE process with extruders already attached is just
needlessly slowed down by this.  And anyone who has managed to get a
1001 error for any reason should probably double check that there isn't
a problem with their lid even if their switch is troggered.