Remove text wrap for print filename fix
BW-5892
http://ultimaker.atlassian.net/browse/BW-5892

You can't wrap text and elide, the TextBody component includes a text wrap by the word, for print status and model filename (where we would prefer the elide) it's necessary to remove the text wrap for it. We don't want to be wrapping the text at all in general.