Remove high temp labs extruder from supported extruders list
The high temp labs is unreleased but we do support it internally at the
machine level. One advantage of this is that the custom temperature list
on the UI can go higher when using this extruder and the test engineering
team sometimes program extruders to this type for quick testing at higher
temperatures without much hassle.

Recently in the wrong extruder popup we updated the logic to be smarter and
actually show the user the supported extruders for a slot if they do happend
to insert an unsupported extruder. The supported list is sent by kaiten and
the UI just converts the internal extruder names to their marketing names
and displays them.

Since the high temp labs extruder remains unreleased we just remove it from
the supported extruder list at the UI level so the user isnt shown this
extruder as a supported one but internal users can still continue using it.

BW-5919
https://ultimaker.atlassian.net/browse/BW-5919