Fixed outdated alert icon
BW-5950
http://ultimaker.atlassian.net/browse/BW-5950

I move the icon to the main menu object which should be easier to keep track if we ever wanted an alert to look different, it appears to only be controlled by the visiblity. If we wanted to add other alerts esp considering we have design goals to implement a notification system.