Download slices from the cloud queue
We never specced out that there would be an API endpoint that you could
use to download the full .makerbot file, but looking into the internal
implementation of teams, there in fact is such an endpoint.  This uses
that endpoint to allow a user to download a file from their cloud queue
into internal storage.  We report download progress and allow the user
to cancel the download if it is being too slow.