Clean up to make automated translation work
BW-5951
http://ultimaker.atlassian.net/browse/BW-5951

The translation tool just refuses to translate any string that has a <
sign in it.  A lot of this comes from having explicit line breaks from
<br> tags in our translation strings.  We completely remove these tags
from translation strings, and we remove as many other html tags as we
can reasonably get away with.

We have two separate uses of <br> here -- a single <br> for a line break
or a double <br> for a paragraph break.  Line breaks are just removed
because we should always be relying on automatic word wrapping (only in
a few cases was word wrapping not already enabled).  For paragraph
breaks we just make each paragraph it's own string to be independently
translated and then add the paragraph breaks in between the translated
strings.

Uses of newline character were also given the same treatment here.  We
don't fail to translate the entire string when these are present but
they do get ignored in the translation.  Some line breaks here are given
the paragraph break treatment since the things that they are separating
are conceptually the same as separate paragraphs.  But I did leave the
line break in "LABS MATERIAL LOADED" because getting word wrap to work
was giving me too much trouble.

All of our other html tags were bold tags around specific parts of text.
Most of these were around things that we are not going to translate,
either a URL or a %1.  For %1 we just pull out the tags into the arg()
for the %1, and for the URLs we just pulled out the entire URL with the
tags around it into a %1 and an arg().  There are still a few cases
where the text in bold was a phrase that we do want to translate and it
would likely be best to translate it within the context of the sentence
it is in, so for these I think I will just manually remove the tags
before translation and then try to put the tags back in place in the
translated text.  (Also a few of these still have their original
translations.)

Also there were "< BACK" strings where I pulled the "< " out of the
translation and a string that was literally only a URL being translated.
And none of the print defect strings had qsTr() on them.