Update the translations via google translate
BW-5951
http://ultimaker.atlassian.net/browse/BW-5951

This includes two lines where we manualy edited out html <b> tags
before the translation and then put these tags back in afterwards, and
also two lines where the old translations were not overwritten because
of the presence of these tags.  And since we left line breaks in LABS
MATERIAL LOADED instead of finding a way to make word wrapping work
there, I added line breaks manually to all of the translations.

(The translation script otherwise overwrote all old translations
because the old translations were all marked as unfinished and I did
not want to write a script to mark all non-empty translations as
finished.)

The arabic translations also seem to have screwed up most of the
instances of %1, so I also fixed those manualy.

And finally, this removes all of the .qm files checked in, because
these files are not even used -- the build system internally generates
it's own .qm files when we build.