Update Filament Drying to use ContentRightSide
We aren't really looking to update the look and feel of the page right
now but ContentRightSide manages to keep the "NEXT" button on screen
throughout all of the screens even when we practically double the amount
of text on these screens by translating it to german.

This does not address the jankiness of the progress bar, which I am
reasonably certain is entirely preexisting, and the images on the left
side do get pushed down very slightly by this change but not in a way
that appears to affect usability.