Protect against bad arguments for FRE steps
BW-5959
http://ultimaker.atlassian.net/browse/BW-5959

We should not be passing invalid step numbers to our FRE tracker or
requesting that it go to the next step from the fre_complete step, but
more importantly we should not segfault if we accidentally do so.  With
this change any attempt to transition to an invalid step will be logged
as an error but otherwise ignored.