Update drop down drawers
This updates the drawers to the new design and hooks them up to the
open/close icon on the right side status icons area in the top bar.
This drawer style and the notification/open/close icon will also be used
for the notifications system though that is not implemted in this commit.

One more change is making the drawers interactive again i.e. swipable
to open/close. This was disabled around Fall 2021 as it was observed
that having the drawers swipable to open from the top edge seemed to
interfere with the touch area of the back button. However, since the close
buttons from all the drawers are removed in the new design swiping up the
drawer to close or clicking a button in the drawer are the only ways to
close it. To
prevent the earlier observed issues the drawers are only interactive in the
open state and not in the closed state. The back button is also hidden when
the drawers are open so this wont affect it in any way.

This commit does not include the notifications system as that is still being
worked out.

BW-5938
https://ultimaker.atlassian.net/browse/BW-5938