Add two simple notifications
Dismissing the extruder calibration prompt popup will post an
informational notification which is dismissable and can be tapped to
get the popup again.

Removing the model extruder or removing filament from the model extruder
will post an error notification which can be dismissed. This alert is
already shown on the material icon in the home screen during this condition
and this notification just mirrors it.

BW-5591
https://ultimaker.atlassian.net/browse/BW-5591