Handle all extruder types
The basic idea here is that getCalibrationPrint will always load a valid
print file, so that the calibration print intro screen can just always
take the user to the start print screen with a valid print file instead
of a start print screen full of garbage, and that whenever possible the
calibration file chosen will be one that could actually print using the
current extruders and materials.  This then allows the existing start
print screen logic to tell the user that they have incompatible material
loaded (though the current language on that screen is not ideal).

This was already acheived for materials because we currently only have
one material combination that we currently support and we hard code that
material combo instead of looking at what is actually loaded.  For
extruders we acheive this by checking if we have a file matching the
model extruder type, but falling back to the composite extruder type if
we don't have a match.  This makes it possible to calibrate using a LABS
extruder (because the printer doesn't care about extruder type mismatch
if you have a LABS extruder) but gives a reasonable error message if you
have a 1A extruder loaded.

This includes a "new" slice for calibration on a 1C extruder for lava,
though this slice is just a copy of the 1XA slice with editted metadata.