Do coarse adjustment regardless of step number
BW-5363
http://ultimaker.atlassian.net/browse/BW-5363

We want the option to retry the process or do auto cal if the second print values still are not within spec. By doing a coarse adjustment regardless when the values are out of spec, the user has the option of retrying but will be closer to finishing the manual z-cal process.