Wait for the calibration print to actually cancel
BW-5986
http://ultimaker.atlassian.net/browse/BW-5986

When we cancel a print, we cannot immediately acknowlege the cancelled
state away because we are not in the cancelled state, we are still in
the cancelling state.  The easiest way around this is to not move back
to the settings pages until we reach the cancelled state.

Specifically, any time that the user uses the X button in the top left
corner to cancel the calibration process and a print is active, we just
ask kaiten to cancel the print and set a flag to watch for when kaiten
is in a state where we could acknowledge the print end state.  We
account for any print end state here because we also trigger this
logic when Xing out of a print that is already in any end state.

This is a little janky because you can X out of the print process
multiple times while cancelling, but this doesn't actually affect
anything.

I also got rid of the onPrintPage property since this was the only
place that we were using it.  Hopefully this doesn't conflict too much
with other tickets...