Revert ts files back to 2.0
BW-5991
http://ultimaker.atlassian.net/browse/BW-5991

These files provide almost no translations for the UI, so the UI is
still mostly in english even when an alternate language is selected.
But we still have not seriously vetted our automatically generated
translations, so for now we would just like to change nothing about
the current state of translations.