Update z offset compensation logic to use last auto cal offset
This switches the entire logic for the z offset compensation feature
to use the last auto cal offset value which is now tracked by kaiten.
The z offset compensation value is now actually computed and displayed
from the difference between the last auto cal offset value and the
current offset value. When changing the compensation value from the
UI the current offsets are directly changed as opposed to tracking the
compensation value as a separate setting. After evey auto cal the
current and the last offsets are set to the same value resetting the
displayed compensation value to 0.

BW-5967
https://ultimaker.atlassian.net/browse/BW-5967