Patch to avoid segfaults on network issues
A printer exhibiting network connectivity issues was observed having a UI crash, and core dump analysis indicated that the issue was that Network::handleRespCheckAuthWithCode was invoked while m_reply was null.  This code does appear to be well written to prevent this from happening under the assumption that a QNetworkReply object can only ever fire a finished signal once.  But I also can't find any guarantee in the qt documentation that it is not possible to fire the finished signal multiple times, and it would appear that this can happen when there are network issues occurring.  So we make sure that we don't segfault when this occurs.