Fixes and improvements to mid-print loading
This fixes two bugs:
1.) Blank screen when trying to load mid-print on M/MX/MXL.
2.) Blank screen when in the preheating step (of standalone loading and
  mid-print loading) on untriggering the extruder switch. The correct
  behavior is to go back to the previous screen telling them to re-insert
  ilament.

This also includes changes to take the user through the entire loading
flow (place desiccant, cut filament tip etc.) even during mid-print
loading for all printers. The challenge is that kaiten starts with
the preheating step when loading mid-print and with the waiting_for_
filament step during normal loading so we use a flag completedNewSpoolSetup
to ensure that the UI doesn't skip right to the preheating screen. However
if the user does trigger the extruder switch while in the the prehetaing
step then they are considered to have finished the new spool setup steps
and will only go back to the screen to prompt them to re-insert filament.

BW-6015
https://ultimaker.atlassian.net/browse/BW-6015