Fix wrong extruder purging after filament jan on model labs extruder
The toolIdx variable which holds the currently loading tool or the intended
to be loaded tool's idx. This fixes a bug where the variable held the last
loaded tool's index and used it for populating the material list when trying
to clear the jam for a labs extruder in the model slot.

BW-6016
https://ultimaker.atlassian.net/browse/BW-6016