Changed image height and width
BW-6012
http://ultimaker.atlassian.net/browse/BW-6012

So I tried a few different things before coming to the conclusion that maybe the height and width here are hardcoded on purpose... Using these values it looks a lot better. And the width in the ThumbnailImage component needs to be the size it is or else it will crash.