Only run material check logic when loading material mid-print
The material check logic can return true or false depending on whether the
check passed or failed respectively and whether the current material in the
bay is allowed to be loaded. In addition it also updates one of the three
flags isMaterialVaid, isUnknownMaterial, isMaterialMismatch which indicate
increasing levels of mismatch. Of these the isMaterialValid flag is used in
the material loading flow for printers with nfc reader (M/MX) to provide the
material valid feedback on the UI and also show the detected material in the
bay and ask the user to insert filament into the slot to go through motor
assisted loading. The isUnknonwMaterial and isMaterialMismatch are both used
to show a mismatching material warning popup and also prevents the user from
loading 3rd party unknown material as well as unsupported materials for that
extruder. These are all only relevant for standalone material on printers with
filament bay (M/MX). For loading in the middle of a print there is no material
check however the isMaterialValid popup is still used in the loading flow during
a print in the same way as described above for the standaolne loading flow.

This commit chnages the conditon to run the material check logic to update the
isMaterialValid flag only during loadingg mid-print and not during unloading.
This is the same behavior as standlone unloading when no material check is run.
Note that even though the material check is run during a mid-print laoding there
are conditons to open the mismatch popup and cancel the loading attempt only
when mismatching during a standlone loading process, so this is effectively a
transparent change that does nothing even if the material check logic sets the
isMaterialMismatch flag when attempting a mid-print load. But this can only
happen if the user replaces a mismatching spool at the end of the loading flow
and attempts to retry loading with the retry button as the load button is disabled
in the material page unless a matching spool as the ongoing print is placed in the
bay. The retry loading/unloading button at the end of loading/unloading flow
currently does not have any logic to be disabled like its counterpart on the
material page.

BW-6015
https://ultimaker.atlassian.net/browse/BW-6015