Updated firmware update related popups
So the popups were not updated when the firmware update process was done, to my knowledge these popups aren't spec'd out... But they come up whenever I update firmware via .zip or if there is an update, so thought it might be nice to have them using our CustomPopup component instead. This makes it a lot cleaner/easier to read (much less code due to using the component instead of building it from scratch). My intention is to change the code here to use our component, not change the functionality in anyway, so it should be working exactly the same.