Update calibrate extruders UI notification posting condition
The extruder change notification is now posted not by clicking the skip
button on the calibrate extruders popup but directly by the same condition
that opens the popup. This makes sense as this is set to be a persistent
notification and the skip button doesn't really fit in the chain of events
as this notification will only go away when kaiten reports that the extruder
are calibrated i.e. basically the same condition that posts this notification
is also what removes it when it is no longer valid.

BW-5591
https://ultimaker.atlassian.net/browse/BW-5591