Encoder feedback flow for loading
BW-5943
http://ultimaker.atlassian.net/browse/BW-5943

Add a new screen to all non assisted loading flows where we wait for the
printer to report that the encoder is moving at the same rate that we
are feeding filament.  This screen has a separate mode for reporting
failure where we then recommend that the user take the lid off of the
machine and directly feed filament into the extruder.

We now use the "loaded_filaments" field from kaiten to determine whether
loading was successful if the user does not give us UI feedback about
whether loading was successful, so we add an explicit field to the bot
model to tell us whether this field is null.  We also add a way for the
UI to tell kaiten when loading is unsuccessful.

This also makes a bunch of changes to the internal flow for loading and
unloading while printing so that we can reuse the same code as when we
are loading and unloading while idle.