Add printer offline notification
BW-6039
http://ultimaker.atlassian.net/browse/BW-6039

Implemented a notification for when the printer is offline by checking if we are not connected to ethernet or wifi.
Added a change to display the wifi symbol instead of ethernet if we are not connected but wifi is enabled.
Added a alert if we are not connected on the Wifi Settings in system settings.