Add hot chamber warning screen to the calibration process
This also simplifies the warning screen component which is now at the
root level of the UI allowing this one instantiation of the component
to be used for all the other processes as well if required. As before
it has two configurations viz. timed check and temperature check for
M/MX and MXL respectively. All of this logic is now contained within
the component.

Note that moving this warning screen to the root level introduces one
behavior change which is, when this warning screen is shown the user
cannot navigate away to a different screen on the UI without dismissing
the warning unlike before.

BW-6069
https://ultimaker.atlassian.net/browse/BW-6069