Use StackLayout instead of SwipeView
BW-6030
https://ultimaker.atlassian.net/browse/BW-6030

This commit replaces the SwipeView containers used throughout the UI with
StackLayout. The StackLayout is very similar to a SwipeView operationally.
It holds its children in an array and has a current index variable which
can be changed to set a particular child as the primary visible element
of the layout. This is also exactly how we were using SwipeViews. However,
the one thing the SwipeViews have is the left/right swiping animation when
changing their current index which we leveraged to visually show the user
a hierarchical view of how the UI is organized. These swiping animations
are fairly long, not customizable and they don't seem to render on-time
(60fps) on the printer leading to a poor experience when using the UI. The
use of StackLayout removes these animations by default

With StackLayout we can specify custom animations for both the previous
element going out of view as well as the new element coming into view, so
we can add the swiping animations if required. This commit adds a very short
fade-in effect for the new screen coming into view which doesn't seem to be
too taxing for the hardware.