Show an error when attempting to start a print with a jammed extruder
This error can be overriden to start a print anyway.

Also renamed the actual start print popup which asks the user to confirm
if the build plate is clear. There is already a popup to ensure that
the build plate is clear that is shown after the print process has started
which was named similarly but that popup is driven by kaiten logic unlike
the UI-only start print popup that only informationally asks the user to
ensure that the build plate is clear before calling the print() kaiten API.

BW-5942
https://ultimaker.atlassian.net/browse/BW-5942