Changes to popups to use BACK and CONFIRM
BW-6034
http://ultimaker.atlassian.net/browse/BW-6034

Changes to standardize how we are displaying information to the user. We want the primary button action to follow along and further confirm the action as a follow-through. Many of the popups required switching the buttons to do the opposite.

Additionally, I re-evaluated what we were trying to portray to the user in some other popups and replacing 'OK' in a few places with 'CONFIRM'. I removed the duplicated aliases out of the CustomPopup component. I moved how we are reading the right and left buttons in the code so it reads left button info above right button info (Left | Right) because it can take some time to figure out which is which.