Remove second button for jam error
BW-6063
http://ultimaker.atlassian.net/browse/BW-6063

We currently allow the user to purge or unload from the jam error screen. Instead we want to direct the user to resolve the process based on whether we are using the assist motors or not. The second button isn't currently being used for any other errors, so I removed it entirely.