Added backIsCancel for nozzle cleaning process
Noticed that we aren't displaying the X (`backIsCancel`) for the Nozzle Cleaning process, but we do have an altBack procedure which opens a cancellation popup when we are in the process. I added the missing variable and set it to display an X during the process/when the cancel popup would be displayed when the back button is pressed.