Correct timezone for Australia central time
BW-6102
http://ultimaker.atlassian.net/browse/BW-6102

I have no idea why Australia/ACT is wrnog but for some reason it is and
has always been.  There are really two different timezones for ACT since
South Australia (Adelaide) uses DST and the Nothern Territories (Darwin)
does not.  Adelaide is significantly bigger so I went with that.