Fix center icon rotation bug
There seems to be a qml bug around changing the rotation property of an
image in a property change. The bug appears to be that the rotation
property isn't actually changing at all. This only seems to affect the
image component and also seems to interact with states and extended states
in an unpredictable way. To get around this, this just adds a dedicated
image component for the center spinning gear instaed of reusing the center
status image component for the spinning gear in the loading state of the
process status icon component.

BW-6052
https://ultimaker.atlassian.net/browse/BW-6052