Show the selected material in top bar during FRE material loading
The issue this fixes is that on XL when loading material during the FRE after
factory resetting an already loaded printer, if that already loaded material
is different from the one being chosen during the FRE loading, the top bar will
show the old material name. Normally when a non-labs extruder is already loaded
the printer won't ask to choose a material when trying to purge/unload (and the
load button is also not visible when already loaded) so a user cannot select a
different material than the one already loaded, but during the FRE for XL we
take the user to an additional instructions screen that then leads to the material
chooser screen which then starts the load process. So in this case on XL a user
can choose a different material than the one already loaded and they will see the
previously loaded material in the top bar for that first load attempt until the
spool journal is updated with the new material at the end of the loading process.
This is not a regression and AFAICT is a bug that has existed since XL launched,
but since at launch-time we probably only tested the FRE loading on new printers
or actually unloaded material and then ran the FRE or the most likley case that
we only had one launch material ABS-R that everyone selected anyway, it wasn't
noticed. I think it is a fairly easy fix conceptually -- we would have to check
the value of the retry_material flag and display it on the top bar. If it not
"None" it means the user just chose the material so we can just show it. If it is
"None" the user didn’t select a material as the printer already knows the loaded
material so we can show that from the spool journal. Currently the retry_material
flag is initialized to "None" when the UI program starts and gets set to the chosen
material when loading/unloading but it isn't being reset at the end of the process,
so we would have to add code to reset it to "None" for this approach and I am a bit
hesitant to make this more generic change as a fix for this issue as this involves
the loading/unloading flow itself. Instead, for this release, I am making the top
bar just show the material name directly from 'retry_material' flag only during the
FRE and stick with the existing logic outside of FRE where the user can never run
into this case with a non-labs extruder.

The labs extruder currently always asks to choose a material during purge/unload and
it can probably be changed at some point to not do that and just use XL style of
using the known loaded material settings (unless we explicitly do not want to do that
for labs for some reason). With only the second fix this bug will still be expressed
if a user chooses a different material during a purge/unload with a labs extrduer.
The first fix will also account for this case.

BW-6015
https://ultimaker.atlassian.net/browse/BW-6105