Set the materialChangeActive flag when loading material in FRE on Method/X
FRE material loading for Method/X uses a dedicated function that has
duplicate code as the regular load() function in MaterialPage.qml. We
earlier added the materialChangeActive flag to prevent unnecessary state
changes in the background when the load/unload process isn't running.
This flag was updated in the regular load/unload function but not here.

Method XL during the FRE material loading just uses the regular load()
function for model extruder loading which is invoked from the material
selector page just like during regular loading and the FRE specific load
function for support extruder loading. We should probably switch this
and Method/X to all use the regular load function during FRE as well.
But for the bugfix for the 2.4 RC we just set this flag manually in the
FRE load function too.

BW-6107
https://ultimaker.atlassian.net/browse/BW-6107