Set FRE complete flag before reset
BW-6104
http://ultimaker.atlassian.net/browse/BW-6104

There is a condition where becuase the FRE step is not being changed until after the reset happens, in the case that a reset is done when the extruders need to be calibrated, the calibration determinant for the popup is triggered. By setting `isFreComplete` to false here, we ensure that a user doesn't see this before reset because the logic for the popup determinant will no longer open the popup.