Update the 'NEXT' button text in FRE material loading flow
While loading material during the FRE on Method the next button at
the end of model material loading leads directly to loading support
material and currently there is no obvious indication that this is
what is happening. On Method XL the user is atleast asked to choose
a material with the support extruder badge displayed above before
being dumped into the loading flow, but on Method/X the landing
screen is an abrupt 'cut material tip' screen after clicking NEXT
though the top bar says Load Material 2. So this commit just updates
the button text to use the same conditon as the onClicked handler
for these buttons to tell the user that clicking the button will
start loading the next material.

We earlier had this button say 'NEXT: LOAD SUPPORT MATERIAL' which
is displayed as 'NEXT: LOAD SUPPORT MAT...', since the button isn't
wide enough. This wasn't too bad TBH atleast with English but product
(Neil) still asked it to be changed back to 'NEXT' which we did, but
when using this it is obvious why we had it the way it was. For this
attempt I used 'NEXT: LOAD MATERIAL 2' which fully fits.

We could also probably update the button in the base FRE step screen
for load material to say 'LOAD MATERIAL 1' for parity. It now says
CONTINUE.