Disable the FRE screen primary button if a process is already running
This seems to have been accidentally modified during the FRE redesign work
before launch. The secondary button also has hardcoded enabled attribute but
it does not have any other explicit conditional logic to control its enabled
property elsewhere like the primary button.

This is also only applicable when in the calibrate and load material step of
the FRE. This was the earlier behavior and I left it as is.

BW-6113
https://ultimaker.atlassian.net/browse/BW-6113