Add a persistent scroll bar to the print failure defects selection list
This print defects list is scrollable but it wasn't obvious that it can
be scrolled and the elements were also spaced far apart that it didn't
readily seem like there were more elements beyond those shown when entering
the page. So there is a chance this might lead users to not provide accurate
feedback.

This commit adds a persistent scroll bar and also adjusts the list spacing
so that it is obvious there are more options for the user to select.