Some UI speed improvements + optimizations
1.) Switch the UI rendering loop from threaded to basic mode. This makes
the UI scene graph render in the main thread instead of using a dedicated
thread. This lowered the frame rate along with significant CPU gains.
Thereaded mode is the default mode and it will generally perform better
especially with multicore CPUs, but it also requires that vsync be enabled
and the display pipeline be properly configured both of which I am not sure
about on our printers. Without these the dedicated scene graph thread can
run at best effort and use a lot more resources without it actually
translating into real life gains. Qt recommends using the basic mode on
systems where configurations cannot be guaranteed. With the threaded mode
the frame rate hovered around 67-68 fps and with the basic mode it hovers
around 60fps sometimes going up to 62 so there is still a slight gain to be
made with enabling vsync. Witht the basic rendering mode the CPU usage also
decreased by 30% from 60% to 40%. I was able to verify that the QSceneGraph
thread wasn't created with the UI spinning up 6 threads now instead of 7.

https://doc.qt.io/qt-5/qtquick-visualcanvas-scenegraph.html#scene-graph-and-rendering

2.) Add an FPS counter item gated behind ui_settings.

3.) Removed the gradient fade just below the top bar. This is a translucent
element with alpha channel that is continuously redrawn anytime anything
moves on the screen like navigating or scrolling through the menus when
elements move behind this translucent item. With the QSG_VISUALIZE debugging
setting this element stands out as a uniquely rendered elemment in every
screen unlike most elements on most screens that are batch rendered. As a
general rule the lesser the number of uniquely rendered elements on the
screen the better.

https://doc.qt.io/qt-5/qtquick-visualcanvas-scenegraph-renderer.html#visualizing-batches

4.) Make the flickable menus overshoot when scrolling when they reach their
boundaries. With the UI and kaiten gains made these are no longer stuttery
and provide a better experience.

5.) Replaced the fade transition when navigating with sliding transitions
like before. These are implemented using a simple custom animator unlike
using the swipeview container's default animations earlier. Again these
seem to render nicely with the new changes.

BW-6135
https://ultimaker.atlassian.net/browse/BW-6135